requirements.txt
fastmcp>=2.12.0
requests>=2.32.0
python-dotenv>=1.1.0
pydantic>=2.12.0
psutil>=7.1.0